only store &mut Connection in SSL ex_data when necessary
We currently need to access the quiche connection state during the TLS
handshake inside BoringSSL's various callbacks, and to do this, when a
Connection structure in created, we store a pointer to it in the SSL
state using the "ex data" API.

This however has the disadvantage that the pointer to the Connection
structure needs to be stable for the whole duration of the connection,
which means that Connection is wrapped in `Pin<Box<...>>` to avoid
applications accidentally moving the structure (thus invalidating the
pointer stored in ex_data).

But we don't actually need the ex data pointer during the whole
connection, in fact, we only need it when calling specific BoringSSL
APIs (e.g. `SSL_do_handshake()`) that in turn call the callbacks
registered by quiche, which means that we can actually set the ex data
pointer just before calling those APIs, and clear it immediately after.

This removes the needs to pin the Connection structure, as it can't be
moved while we hold a mutable reference it (due to being inside one of
quiche's methods), and simplifies the API a bit (as applications don't
need the `Pin<Box<...>>` wrapper anymore).

Fixes #1210.