h3: add HTTP/3 datagram RFC version
With this change, we add support for advertising the HTTP/3 datagram SETTING
value (0x33) defined in draft 11 and RFC 9297 when an h3 Connection is
configured to have datagrams enabled.

The format of the HTTP/3 datagrams are pretty much structurally identical,
so there is not much meaningful difference between versions. Applications that
rely on HTTP/3 datagrams, such as to use CONNECT-UDP, would apply their own
semantics to these datagrams and can manage that themselves. Therefore, this
change doesn't provide fine grain control over versions.