h3: avoid needless copy when checking dgram queue
In an old version of the DATAGRAM poll event, it returned an HTTP/3 datagram
flow ID, which required peeking at the first 8 bytes of the QUIC frame
payload at the front of the queue.

When we changed to edge-triggered events, we changed the logic to
always return the placeholder value 0. However, we didn't change
the logic to avoid peeking at the data, which meant we left a
needless 8-byte allocation and copy.

This change removes the needless copy, replacing it instead with a check
of the queue length to determine if the edge-trigger should fire
or not.