lib: prevent invalid unidirectional stream_shutdown
Previously, quiche would allow a sending endpoint to try and STOP_SENDING a locally-initiated stream, or RESET_STREAM a remotely-initiated stream. Both of these actions are semantically wrong, and prohibited by RFC 9000. On the receive side, quiche enforces these checks, meaning it is easy for an application to do the wrong thing and cause a connection to be terminated by the remote peer.

With this change, we add sender-side checks to prevent invalid behavior and instead return an error code.