fix server app close before 1RTT never being sent
If a server does an app close after a connection is established but
before it has sent the Handshake ack the close will not be sent. (app
close error codes cannot be sent during the Handshake)

This fix checks if the current packet type is Handshake and ignores
is_closing if the local_error is an app error.

The resulting behaviour is that the Handshake ack packet will be sent
followed by a 1RTT with the connection_close.

Fixes #1118.