packet: fix hard-coding constants
Previously, both MAX_PKT_NUM_LEN and SAMPLE_LEN are hard-coded at `encrypt_hdr` while it is not the case at `decrypt_hdr`. If any of the constant is changed in the future, one must carefully sync up the redundant hard-coded constant at `encrypt_hdr`. After the change, it is more consistent and the hard-coding is alleviated.

No functional code is changed.