ranges: add an inline version of RangeSet
Currently the RangeSet is always backed by a BTreeMap, and the map
is allocated every Ack packet. This change is not unlike SmallVec
for Vec, where instead of allocating, a limited number of ranges can
be stored inline. Because in practice Ack packets will rarely ack
more than one or two ranges, this avoids a significant number of
allocations.

RangeSet is also used by the SendBuf to track the numbers of acked
packets. Although in this use case allocations are already rare,
with inline RangeSet the insertion operation is now also significantly
faster.