reduce number of path lookups in the hot path on send
During send_single, the current path is retrieved many times over
which is both not free from a performance PoV and also not very
readable. This change makes sure paths.get_mut is only called once
per packet, which requires some workaround for the borrow rules.
Overall the result is a cleaner code with slightly improved perforamnce.

In addition this change also keeps a single reference to the current
PktNumSpace, which is purely a readabity improvement.