downgrade error epoch to Initial if remote peer can't decrypt yet
Motivation:

We should downgrade the epoch to Initial if the remote peer cant decrypt yet, failing to do so might cause the remote peer to drop the frames on the floor and trying to resend.

Modifications:

Check if we need to downgrade to Intial and if so do it.

Result:

Remote peer can always handle frame