mark streams as writable when blocked by connection capacity
When using `stream_writable_next()` a stream might initially be
writable, but then become blocked due to connection-level send capacity.

When this happens the stream never ends up being marked writable again,
even once the capacity increases.

Instead, when a write is blocked by connection capacity, mark the stream
as writable again immediately, so when the capacity increases it can be
returned by `stream_writable_next()` again later.

This issue was masked in some cases by a slight bug that would mark a
streams as writable every time it was looked up in the connection's
streams map, while the intention was to do this only for streams that
are newly created (which is why the test changes from d1172f46 didn't
catch this earlier).