tls: fix a pile of incorrect bindings
See issue #1559. This does not fix the crate's unsafe bindings strategy,
but corrects some of the mistakes that snuck in as a result of this
unsafe strategy.

One thing to call out is that sk_num and sk_value's index types were
wrong. This would have caused a problem were the crate not saved by most
64-bit calling conventions defining the upper half of the register when
a 64-bit register is used to pass a 32-bit value. (But I have not
carefully checked all calling conventions for this. The real fix is to
stop writing bindings this way. This bindings strategy is even less safe
than calling the library from C, which is quite a low bar to fail!)

Beyond that, the issues are mostly const-ness, though one notable issue
is that function pointers need to be imported as Option. Also the return
value for set_min_proto_version and set_max_proto_version was being
silently ignored.