pass a SslContextBuilder instead of SslContext
quiche currently unwraps the SslContextBuilder and modifies it
with the boring FFI functions directly. This assumes the SSL
context is exclusively owned by quiche, but that may not be
the case as SslContext values are clonable. The function should
take a SslContextBuilder instead, which guarantees that the
inner SSL context is not shared with anything else prior to
calling SslContextBuilder::build.