lib: skip dgram frame size calc when none to send
Previously, we calculated the value of the largest DATAGRAM frame that
could be sent every time we wanted to send a packet. The calculation is
controlled by peer TPs, which can be sent even if the peer doesn't
really want us to use the DATAGRAM on the connection (e.g. conventional
HTTP/3 serving).

We only need the DATAGRAM size when we have datagrams to actually send.
This change implements a check of the internal DATAGRAM send queue to
avoid the calculation.