h3: if readable after Finished, call recv again
If an H3 stream gets into a Data state that is not cleared for the
stream (for example because downward backpressure), and a ResetStream
arrives, the poll will transition the stream to finished and never clear
the readable state of the stream, it will forever stuck in the readable
state. This will mean that poll will return Finished and not Reset for
the stream, and will have a somewhat significant performance impact,
as for long lived connections it will take longer and longer to poll
before the iterator reaches a stream that is actually readable.