h3: avoid unnecessary work on control streams
Previously, quiche's poll() function would attempt to do work on control or
QPACK streams, whether they were readable or not. Since we encourage apps to
call poll() after each read loop, this could result in wasted work.

With this change, we'll return early if the stream is not actually readable.