crypto: replace ring AEAD API for header protection
This is currently only used for generating header protection masks,
which, again, doesn't seem like something worth pulling a whole
dependency for. This may or may not fall under FIPS scope.

Note that the ring dependency itself is still needed to build examples,
but it can be downgraded to `dev-dependencies`.