lib: change qlog connection_closed trigger value
Previously, it was common for quiche to log a connection_closed
event with a trigger value of "error". This was caused by some
simple checking of peer_error or local_error inside of the
quiche library when closing the connection.

This change tweaks the logic in mark_closed() just a little. If
the error was at the transport-level with the QUIC NO_ERROR code,
we'll mark it as "clean". Any other transport code remains
as "error". If the error was at the application-level, we are
not application protocol-aware and can't make any reliable guess
about what a code means, so it it better to omit the trigger.