crypto/boringssl: perform dummy seal when client resets keys too
This is yet more fallout from cb220a59 (also see 393f1cc). This time the
missing dummy seal is for clients, after the initial epoch is reset
(e.g. due to retry or version negotiation), which also causes the
initial key material to be dervied again.

When that happens, and if the client's PTO fires (e.g. because the
server is slow to respond) that will trigger the client to create a new
Initial packet that will fail to seal because the initial seal context
expects a packet number of 0, while the actual packet number is greater
than that as it's a retransmission.

This only happens when the client's initial PTO fires because otherwise
it would just send a Handshake packet instead, without using the initial
key material.