tokio-quiche: Add error log for failed packet router
So far, server-side `InitialPacketRouter`s that stopped with an error
(from the underlying UDP socket) closed without any notice. The only way
to observe this is through the `QuicConnectionStream` closing. To help
with debugging, this change adds an error log similar to the one we
already have for client-side connections.