octets: add bounds check in unsafe code in macros
The `put_u` and `peek_u` macros in octets contains unsafe blocks. This
commit adds a safety check to make sure the length we are being passed is
within the proper bounds. Without this check it's possible to access
invalid memory.

Note that the macro is not exported and all current uses of it are
safe. So this commits just prevents future issues.