tokio-quiche: add support for setting per-response h3 priorities
Currently tokio-quiche always sets the same default ptiority for all h3
responses. This change adds support for applications to specify an
optional priority on a response-by-response basis instead, while still
keeping the current default priority for backwards compatibility.

This is done by adding a new `Option<Priority>` field to
`OutboundFrame::Headers()`. Applications that don't want to provide any
priority can simply set this to `None`, in which case the existing
default priority logic applies.

In order to allow setting priorities for additional headers as well, a
new method `send_additional_headers_with_priority()` is also added. This
is required because the `Priority` fields are private, which means it
would not be possible to call the lower level `stream_priority()` method
with their values.

Finally, `Clone` and `Copy` traits are derived for `Priority` as well to
simplify use in applications.