tokio-quiche: Add input path to file errors (#2046)
BoringSSL does this internally when we ask it to load a PEM file, but
for RPK we call `fs::read` ourselves. The error message is not very
helpful in those cases, as we simply bubble up an io::Error without any
context.

`anyhow` is already an implicit dependency via foundations, so I figured
using it was fine.