h3i: don't busy-loop on wait_for_data if else arm fires
Turns out that the `else` arm returning nothing can end up starving the
IOW sleep arm, leading to hung tests. To avoid that, and also avoid
panicking due to a lack of an `else` arm, we now just sleep forever if
we have nothing to do. This allows the IOW to drive itself.