h3: move Huffman encoding / decoding to octets crate
Huffman encoding is used by HPACK as well, so it makes sense to share
the implementation rather than duplicate all this code.

This also somewhat simplifies the decoding code which was split across
multiple functions and structures for not much benefit (if anything it
made the code harder to follow).