task-killswitch: Use AbortHandle to reduce drop cost
JoinHandle is expensive to drop as it allows awaiting a task and
accessing its output. By keeping it around in the task map, we
centralize this drop cost in a single thread with cold memory.

An AbortHandle instead only keeps around a cheap(er) reference to the
task. The JoinHandle is dropped almost immediately after the task is
spawned, while it is still fresh in memory/caches. Tokio can take
advantage of the dropped JoinHandle by avoiding unncessary work.

(Alternatively, this would also allow us to return the JoinHandle to the
caller.)