packet: use longer packet number length encoding when necessary
The decoding method described in RFC 9000 treats the encoded packet number as a signed integer. We need to encode the sign bit to ensure correct decoding when 128+ packets are lost.

This PR fixes the packet number length computation, and adds a test to verify that the returned packet length is sufficient to decode the next packet number assuming a loss of 1..254 packets.

Fixes #2180.